Move bulk edit helpers out of product edit utils
